refactor: move DatapageHelpers into baker
