--- datapage/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../devTools/tsconfigs/tsconfig.base.json",
-    "compilerOptions": {
-        "outDir": "../itsJustJavascript/datapage",
-        "rootDir": "."
-    },
-    "references": [
-        { "path": "../gitCms" },
-        { "path": "../explorer" },
-        { "path": "../db" }
-    ]
-}
