+++ tsconfig.json
@@ -17,9 +17,6 @@
             "path": "./explorer"
         },
         {
-            "path": "./datapage"
-        },
-        {
             "path": "./site"
         },
         {
